feat: migrate layout/singleColumnCentered and BaseLayout component (#1292)
* feat: migrate BaseLayout and SingleColumnCenteredLayout

---------
