+++ src/components/layout/index.ts
@@ -1,1 +1,2 @@
 export { repeatArea } from './BaseGrid';
+export { SingleColumnCenteredLayout } from './SingleColumnCentered';
